Fix duplicate enrich column bug
When running SARIMAX, which inherits from ARIMA, enrich was being called
twice - one in the fit method of SARIMAX and another in the super call
to ARIMA's fit. As we're doing a concat, the default behaviour of pandas
is to leave both columns present.

Removing the extra enrich call in SARIMAX fixed the issue, adding the
verify_integrity flag True in the concat makes this easier to find in
future.